refactored test functions into their respective utils folders
