+++ contracts/market/CreateLoanFacet.sol
@@ -261,14 +261,16 @@ library CreateLoanLib {
      * @return id_ the new ID requested, which stores it in the loan data
      */
     function newID() internal returns (uint256 id_) {
-        Counters.Counter storage counter = MarketStorageLib.store()
+        Counters.Counter storage counter = MarketStorageLib
+            .store()
             .loanIDCounter;
         id_ = Counters.current(counter);
         Counters.increment(counter);
     }
 
     function currentID() internal view returns (uint256 id_) {
-        Counters.Counter storage counter = MarketStorageLib.store()
+        Counters.Counter storage counter = MarketStorageLib
+            .store()
             .loanIDCounter;
         id_ = Counters.current(counter);
     }
