+++ contracts/market/RepayFacet.sol
@@ -353,7 +353,8 @@ library RepayLib {
 
         // Calculate available collateral for reward
         if (collateralAmount > 0) {
-            collateralValue_ = AppStorageLib.store()
+            collateralValue_ = AppStorageLib
+                .store()
                 .priceAggregator
                 .getValueFor(
                     LibLoans.loan(loanID).collateralToken,
@@ -399,7 +400,8 @@ library RepayLib {
         // if the lending reward is less than the collateral lending tokens, then aggregate
         // the value for the lending token with the collateral token and send it to the liquidator
         if (rewardInLending <= collateralInLending) {
-            uint256 rewardInCollateral = AppStorageLib.store()
+            uint256 rewardInCollateral = AppStorageLib
+                .store()
                 .priceAggregator
                 .getValueFor(
                     LibLoans.loan(loanID).lendingToken,
@@ -432,7 +434,8 @@ library RepayLib {
         address recipient,
         uint256 value
     ) private {
-        EnumerableSet.AddressSet storage tokens = MarketStorageLib.store()
+        EnumerableSet.AddressSet storage tokens = MarketStorageLib
+            .store()
             .escrowTokens[loanID];
         uint256 valueLeftToTransfer = value;
 
@@ -482,7 +485,8 @@ library RepayLib {
             if (token == LibLoans.loan(loanID).lendingToken) {
                 balanceInLending = balance;
             } else {
-                balanceInLending = AppStorageLib.store()
+                balanceInLending = AppStorageLib
+                    .store()
                     .priceAggregator
                     .getValueFor(
                         token,
