+++ contracts/market/libraries/LibCollateral.sol
@@ -105,7 +105,8 @@ library LibCollateral {
         // Check if collateral escrow exists
         if (address(e(token)) == address(0)) {
             // Create escrow
-            address escrow = AppStorageLib.store()
+            address escrow = AppStorageLib
+                .store()
                 .collateralEscrowBeacon
                 .cloneProxy("");
             ICollateralEscrow(escrow).init(
