+++ contracts/shared/libraries/LibDiamond.sol
@@ -314,8 +314,8 @@ library LibDiamond {
                 .facetFunctionSelectors[_facetAddress]
                 .functionSelectors[lastSelectorPosition];
             ds.facetFunctionSelectors[_facetAddress].functionSelectors[
-                selectorPosition
-            ] = lastSelector;
+                    selectorPosition
+                ] = lastSelector;
             ds
                 .selectorToFacetAndPosition[lastSelector]
                 .functionSelectorPosition = uint16(selectorPosition);
